requirements.txt
# Core ML/Data
torch==2.1.2
transformers==4.37.2
pandas==2.1.4
polars==0.19.19
scikit-learn==1.3.2
numpy==1.24.3

# API & Web
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
python-multipart==0.0.6

# Orchestration
apache-airflow==2.8.0
apache-airflow-providers-kubernetes==8.5.0
apache-airflow-providers-postgres==5.10.0

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# OAuth & Security
python-keycloak==3.8.0
authlib==1.3.0
PyJWT==2.8.1

# Data Validation & Versioning
great-expectations==0.18.13
dvc==3.48.4
dvc-s3==3.2.0  # Optional for S3 backend

# Monitoring & Logging
python-json-logger==2.0.7

# Testing
pytest==7.4.4
pytest-asyncio==0.23.2
pytest-cov==4.1.0
httpx==0.25.2

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.0
requests==2.31.0
